putting all tests in their own folders and organizing
