+++ tests/texting/test1.py
@@ -1,10 +1,10 @@
 from src import lynxy
 from rich import print
 
-inst = lynxy.Lynxy(('', 11112), True)
+inst = lynxy.Lynxy(('', 11111), True)
 print('initialized')
 host = inst.get_host()
-target = ('192.168.68.130', 11111)
+target = ('192.168.68.124', 11112)
 
 print(f'host: {host}')
 print(f'target: {target}')
@@ -12,10 +12,10 @@ print(f'target: {target}')
 inst.connect(target)
 print('connected')
 
+
 @inst.event(lynxy.Constants.Event.ON_MESSAGE)
 def recv(msg: lynxy.Pool.Message): 
-    print(f'{msg.created_at} -> {msg.recieved_at}')
-    # print(f'{msg.created_at} -> {msg.recieved_at}: {msg.content}')
+    print(msg.content)
 
 @inst.event(lynxy.Constants.Event.ON_CLOSE)
 def close(msg):
@@ -27,4 +27,18 @@ while True:
     if msg == 'exit': 
         inst.close()
         break
-    inst.send(msg)
+    inst._comm._send(msg)
+
+
+## uncomment for sending bee movie, comming out above while loop
+
+# # https://courses.cs.washington.edu/courses/cse163/20wi/files/lectures/L04/bee-movie.txt
+# with open(r'D:\VScode\packages\lynxy\bee.txt', 'r') as f:
+#     contents = f.read()
+# print(len(contents))
+# inst.send(contents)
+
+# while True:
+#     input('-> ')
+#     break
+# inst.close()
